scripts/requirements-load.txt
# For load_marathon_export.py (OS-level load with direct DB connection)
psycopg2-binary>=2.9
ijson>=3.2
